chore: move referenceArea atom into shared_state
