+++ src/core/shared_state/referenceArea.ts
